+++ src/classes/Application/API/Admin/APIScreenRights.php
@@ -22,4 +22,7 @@ class APIScreenRights
     public const string SCREEN_API_KEYS_SETTINGS = APIRightsInterface::RIGHT_VIEW_API_CLIENTS;
     public const string SCREEN_API_KEYS_SETTINGS_EDIT = APIRightsInterface::RIGHT_EDIT_API_CLIENTS;
     public const string SCREEN_API_KEYS_LIST = APIRightsInterface::RIGHT_VIEW_API_CLIENTS;
+    public const string SCREEN_API_KEYS_METHODS = APIRightsInterface::RIGHT_VIEW_API_CLIENTS;
+    public const string SCREEN_API_KEYS_METHODS_EDIT = APIRightsInterface::RIGHT_EDIT_API_CLIENTS;
+    public const string SCREEN_API_KEYS_RIGHTS = APIRightsInterface::RIGHT_VIEW_API_CLIENTS;
 }